backend/requirements.txt
flask
flask-cors
python-dotenv
supabase
pyjwt
openai
requests
python-dateutil
